Create proca-cli, rename project to proca_cli
