+++ package.json
@@ -1,17 +1,18 @@
 {
-  "name": "proca-gw",
-  "version": "1.0.0",
-  "description": "client for proca",
-  "main": "build/index.js",
+  "name": "proca_cli",
+  "version": "1.1.0",
+  "description": "cli[ent] for Proca",
+  "main": "dist/index.js",
   "repository": "git@github.com:TechToThePeople/proca-gw.git",
   "author": "xavier.dutoit <xavier@tttp.eu>",
   "license": "aGPL3",
   "bin": {
-    "proca-gw": "bin/proca-gw"
+    "proca-cli": "bin/proca-cli"
   },
   "scripts": {
     "install": "npm run build",
-    "build": "npx parcel src/index.js --target=node"
+    "build": "npx parcel build src/index.js --target=node",
+    "watch": "npx parcel src/index.js --target=node"
   },
   "dependencies": {
     "base64url": "^3.0.1",
